+++ packages/shared/shared-types.d.ts
@@ -61,7 +61,8 @@ export interface DesktopSettingsType {
   syncAllAccounts: boolean
   /** @deprecated The last used file location for the save dialog is now only kept in memory and not persisted anymore between sessions. */
   lastSaveDialogLocation: string | undefined
-  experimentalEnableMarkdownInMessages: boolean
+  /** @deprecated */
+  experimentalEnableMarkdownInMessages?: boolean
   enableWebxdcDevTools: boolean
   /** set to false to disable the confirmation dialog for loading remote content */
   HTMLEmailAskForRemoteLoadingConfirmation: boolean